requirements.txt
aniso8601==8.0.0
astroid==2.3.3
attrs==19.3.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
flask-restplus==0.13.0
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-vision==1.0.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.1
pandas==1.0.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.4.4
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2
zipp==3.0.0
